Move functions to credential package
