+++ pkg/credential/error.go
@@ -1,0 +1,9 @@
+package credential
+
+import (
+	"github.com/giantswarm/microerror"
+)
+
+var missingValueError = &microerror.Error{
+	Kind: "missingValueError",
+}
